requirements.txt
suno-bark @ git+https://github.com/anisingh1/bark-with-voice-clone.git
scipy
nltk
uvicorn
fastapi
pydantic
openai-whisper
torch
torchaudio
pandas